--- a/v3/go.mod
+++ b/v3/go.mod
@@ -3,6 +3,7 @@ module github.com/igm/sockjs-go/v3
 go 1.14
 
 require (
+	github.com/gorilla/mux v1.7.4
 	github.com/gorilla/websocket v1.4.2
 	github.com/stretchr/testify v1.5.1
 )