Relates to #81 

@igm Just want to validate whether this is a desired change?

Notes:
1) I have not considered alternatives to `gorilla/mux` since it's popular, robust and looks nice near `gorilla/websocket`
2) I have not measured performance difference yet
3) Originally I removed requirement to pass `prefix` to SockJS handler constructor making user to use `http.StripPrefix` from outside, but then I saw that you don't want to break API globally in v3 so put prefix back to constructor