@FZambia 2020-05-23
Added `BenchmarkServe` and compared with master, unfortunately results are not very good:
```
fz@sockjs: benchstat old.txt new.txt
name     old time/op    new time/op    delta
Serve-4    1.73µs ± 2%    3.30µs ± 1%   +90.58%  (p=0.000 n=9+10)

name     old alloc/op   new alloc/op   delta
Serve-4      166B ± 3%     1442B ± 0%  +771.00%  (p=0.000 n=10+8)

name     old allocs/op  new allocs/op  delta
Serve-4      7.00 ± 0%     16.00 ± 0%  +128.57%  (p=0.000 n=10+10)
```