@FZambia 2020-05-23
Also experimented a bit with https://github.com/julienschmidt/httprouter - it gives better results:
```
fz@sockjs: benchstat old.txt newest.txt
name     old time/op    new time/op    delta
Serve-4    1.73µs ± 2%    1.22µs ± 4%   -29.25%  (p=0.000 n=9+9)

name     old alloc/op   new alloc/op   delta
Serve-4      166B ± 3%      663B ± 2%  +300.18%  (p=0.000 n=10+10)

name     old allocs/op  new allocs/op  delta
Serve-4      7.00 ± 0%     11.00 ± 0%   +57.14%  (p=0.000 n=10+10)
```

Faster in terms of raw operation count, but allocates a bit more. Code in separate branch: https://github.com/igm/sockjs-go/tree/httprouter, it required to handle some paths manually due to httprouter limitations and still needs some additional work for iframe, I just stopped on some point - not sure this all worth it - we can fix #81 simply modifying current mapping implementation a bit.