@igm 2020-05-23
@FZambia  thank you for thorough testing and benchmarking both routers, interesting results. I'll look at it more in detail.

Regarding #81 , is it still the issue in `v3` by your opinion? I thought it works fine in `v3` already. I've written a test for those cases https://github.com/igm/sockjs-go/commit/2a321445b0f6f59d66ff59322bbceb0880cad395 and they pass in `v3`.