@FZambia 2020-05-23
> Regarding #81 , is it still the issue in v3 by your opinion? 

Actually I missed the fact that you already fixed that in v3:)

I suppose then there is no sense to move to `mux` since it adds some overhead and no real profit for users, moving to `httprouter` could provide some performance benefit, but not too radical to go with it without additional reason.